Added tests for editing mood event
